requirements.txt
gib
gib_bot
discord.py
sqlalchemy
